# perf(winston): marginally faster log record building